Support Observables (#10)
When the server first encounters an `Observable`, it will `subscribe` to
it and send updates to the client for every notification. The client
will then use these updates to set properties on the client's object
which can then be read synchronously. Furthermore the caller can call
the client object's `subscribe` method to get notified of updates.